Rename GameComponents folder to Game.
