+++ src/components/pages/GameContainer/Game/index.ts
